make the prompts prettier
