+++ src/components/NewsFeed/Prompts/PromptBgImage.js
@@ -1,0 +1,149 @@
+import './Prompt.scss'
+
+const PromptBgImage = () => {
+    return ( 
+        <div className='prompt-background-image'>
+            <div class='scene'>
+                <div class='clouds'>
+                <div class='cloud cloud-3'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                </div>
+                <div class='cloud cloud-1'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                </div>
+                <div class='cloud cloud-2'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                </div>
+                <div class='cloud-alt'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                </div>
+                </div>
+                <div class='sky'>
+
+                <div class='sun'></div>
+
+                <div class='sky-1'></div>
+                <div class='sky-2'></div>
+                <div class='sky-3'></div>
+
+                <div class='background-clouds'>
+                    <div class='cloud-1'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                    </div>
+                    <div class='cloud-2'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                    </div>
+                    <div class='cloud-3'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                    </div>
+                    <div class='cloud-4'>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bubble'></div>
+                    <div class='bottom'></div>
+                    </div>
+                </div>
+                </div>
+
+                <div class='mountains'>
+                <div class='mountain'></div>
+                <div class='mountain-2'></div>
+                <div class='mountain-3'></div>
+                <div class='mountain-4'></div>
+                <div class='shadow'></div>
+                </div>
+
+                <div class='bottom-section'>
+
+                <div class='objects-foreground'>
+                    <div class='house-2'>
+                    <div class='roof'>
+                        <div class='front'></div>
+                        <div class='side'></div>
+                    </div>
+                    <div class='main'>
+                        <div class='front'>
+                        <div class='window'></div>
+                        </div>
+                        <div class='side'>
+                        <div class='window'></div>
+                        </div>
+                    </div>
+                    <div class='light'></div>
+                    </div>
+                </div>
+
+                <div class='ground'>
+                    <div class='hill-bottom'>
+                    </div>
+                    <div class='ground-bg'></div>
+                    <div class='ground-main'>
+                    <div class='light'></div>
+                    </div>
+                    <div class='hill'></div>
+                    <div class='trees'>
+                    <div class='tree'></div>
+                    <div class='tree'></div>
+                    <div class='tree'></div>
+                    <div class='tree'></div>
+                    <div class='tree'></div>
+                    <div class='tree'></div>
+                    </div>
+                    <div class='trees-2'>
+                    <div class='tree'></div>
+                    <div class='tree'></div>
+                    <div class='tree'></div>
+                    </div>
+                    </div>
+                    </div>
+                </div>
+            </div>
+     );
+}
+
+export default PromptBgImage;
